refactor: share schema-echo handling for prompt refinement (#6794)
