+++ server/services/universeBuilderRefine.test.js
@@ -3,74 +3,11 @@ import { __testing } from "./universeBuilderRefine.js";
 import { mergeInfluencesWithLocksAdditive } from "./universeBuilder.js";
 
 const {
-  extractRefinementJson,
   buildWorldRefinePrompt,
   collapseStyleDirectionDupes,
   mergeCategoriesWithLocks,
   mergeCompositesWithLocks,
 } = __testing;
-
-describe("universeBuilderRefine.extractRefinementJson", () => {
-  it("parses a raw refinement object", () => {
-    const obj = {
-      starterPrompt: "a darker scavenger universe",
-      stylePrompt: "gritty palette, deep shadows",
-      negativePrompt: "cute, neon",
-      rationale: "pushed mood toward grim",
-    };
-    expect(extractRefinementJson(JSON.stringify(obj))).toEqual(obj);
-  });
-
-  it("strips ```json fences", () => {
-    const fenced =
-      '```json\n{"starterPrompt":"x","stylePrompt":"y","negativePrompt":""}\n```';
-    expect(extractRefinementJson(fenced)).toMatchObject({
-      starterPrompt: "x",
-      stylePrompt: "y",
-    });
-  });
-
-  it("skips preamble before the JSON", () => {
-    const raw =
-      'Here is the refinement:\n{"starterPrompt":"x","stylePrompt":"y"}\nend';
-    expect(extractRefinementJson(raw)).toMatchObject({ starterPrompt: "x" });
-  });
-
-  it("skips a schema-example block that has a <…> placeholder starterPrompt and parses the real block (Codex CLI prompt echo)", () => {
-    const raw = [
-      "codex banner",
-      // The prompt template body — its first balanced { ... } block contains
-      // <…> placeholders that walked past extractRefinementJson by mistake
-      // would surface as "AI returned schema placeholder" instead of finding
-      // the real response below it.
-      '{"starterPrompt":"<full rewritten…>","stylePrompt":"<…>","negativePrompt":"<…>"}',
-      "codex response:",
-      '{"starterPrompt":"a darker universe","stylePrompt":"gritty","negativePrompt":""}',
-    ].join("\n");
-    const out = extractRefinementJson(raw);
-    expect(out.starterPrompt).toBe("a darker universe");
-    expect(out.stylePrompt).toBe("gritty");
-  });
-
-  it("throws when ONLY schema-placeholder blocks are present", () => {
-    const raw = '{"starterPrompt":"<placeholder>","stylePrompt":"<x>"}';
-    expect(() => extractRefinementJson(raw)).toThrow(/schema placeholder/);
-  });
-
-  it("throws on empty / non-string input", () => {
-    expect(() => extractRefinementJson("")).toThrow(/Empty AI response/);
-    expect(() => extractRefinementJson(null)).toThrow(/Empty AI response/);
-  });
-
-  it("throws when no balanced JSON object with starterPrompt is present", () => {
-    expect(() => extractRefinementJson("just prose, no json")).toThrow(
-      /Invalid JSON/,
-    );
-    expect(() => extractRefinementJson('{"prompt":"unrelated shape"}')).toThrow(
-      /Invalid JSON/,
-    );
-  });
-});
 
 describe("universeBuilderRefine.buildWorldRefinePrompt", () => {
   it("includes originals + influences + feedback verbatim", () => {
